Create build scripts that are compatible with github actions (and others)
We create some simple scripts that aim to support a reasonable CD process.

* On each push to the master branch, we push every function at that version.

* On each tag, we match that to a function directory and push that to
  the container repo.  As with kpt functions, the format is:
  `<directory path>/<version>`

The goal is that github actions should bind variables like IMAGE_REPO
and GIT_REF to the correct version, but otherwise we try to minimize
the amount of work we do in github actions.  Most of the logic should
be in a script in `dev/cd/`, that clearly corresponds to the trigger
(e.g. `after-branch-push` for pushes to the main branch,
`after-tag-push` for when a tag is pushed).

If we need to run in other CI/CD systems, it should suffice to hook up
the trigger to the same scripts under `dev/cd/`.